very crude fix for table width formatting. relocate filter row out of table
This is a super-crude fix (my apologies) but it works for me. I'm facing an issue with formatting of my table width. In order to fix this, I need to style my table with table-format:fixed. When I do this, the column widths are all taken from the first row of the table. Because I have Search enabled, the first row of the table is the Search row, and so my column styling is all ignored. This fix wraps the table in a div, then moves the Search row up a level so it's in the div, but outside of the table. Now my column width formatting is being correctly applied and I can style my table correctly.