Fix for #167
A likely quickfix would be to select the first page only when the number of pages changes. 
Although this will not reset the page chosen whenever the modified data has the same number of pages as the safe copy, it would always display the table, then the user can reset the pages manually if desired.