Add multi select function using Shift and Ctrl
Add another option for the selection that allows multiple lines to be
selected using the Shift and Ctrl keys in the same way as in Windows.
I don't have a full development environment available to me on my
corporate workstation.  So this was developed by modifying the full
release smart-table-1.4.5.js file.  Then the changes were copied into
the individual source files.

I have tested this in IE9 and Chrome 36.0.1985.125.

You will probably want to do more testing than I am able to before
including it in the release.