Added a $watch to the first item in safeSrc
Added a `$watch` to the first item in the `stSafeSrc` array. This change `$watch`'s the first element of the `stSafeSrc` if it exists and updates the `stTable` accordingly.

Although #205 addresses a similar issue it doesn't address the case of many similar length arrays being swapped out of an instance of `stTable`.